pyproject.toml
[tool.poetry]
name = "joinup"
version = "0.1.0"
description = ""
authors = ["ktowen <pedroall21@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
celery = {extras = ["redis"], version = "^5.3.6"}
django = "^5.0.3"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^7.3.0"}
djangorestframework = "^3.14.0"
python = "^3.12"


[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"
ipython = "^8.22.2"
ruff = "^0.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
